# dental-erp/api/openapi.yaml
openapi: 3.1.0
info:
  title: DentaCore API
  version: 1.0.0
  license: { name: Proprietary, identifier: LicenseRef-Proprietary }
  summary: Dental Clinic ERP - Faza 1 (MVP) public API
  description: |
    Bütün endpoint-lər API Gateway üzərindən `/v1` prefiksi ilə açılır.
    Tenant subdomain-dən (`klinika.dentacore.app`) və ya JWT `tid` claim-dən təyin olunur.
    Xətalar RFC 9457 (`application/problem+json`). Siyahılar cursor pagination ilə.
    Yazma əməliyyatlarında optimistic concurrency: `If-Match: <rowVersion>` (uyğunsuzluq → 412).
    Pul məbləğləri string deyil, `number` (2 onluq) və ayrıca `currency`.
servers:
  - url: https://{tenant}.dentacore.app/v1
    variables: { tenant: { default: demo } }
  - url: https://clinic.local/v1
    description: Local server
security:
  - bearerAuth: []
...
paths:
  /auth/login:
    post:
      operationId: postAuthLogin
      tags: [Auth]
      summary: Giriş (email + parol)
      security: []
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/LoginRequest' } } }
      responses:
        '200': { description: Tokenlər, content: { application/json: { schema: { $ref: '#/components/schemas/TokenResponse' } } } }
        '202': { description: 2FA tələb olunur, content: { application/json: { schema: { $ref: '#/components/schemas/TwoFactorChallenge' } } } }
        '401': { $ref: '#/components/responses/Problem' }
        '423': { description: Hesab müvəqqəti bloklanıb, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
        '429': { $ref: '#/components/responses/Problem' }
...
  /auth/refresh:
    post:
      operationId: postAuthRefresh
      tags: [Auth]
      summary: Refresh token rotasiyası (reuse aşkarlanarsa bütün family ləğv olunur)
      security: []
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [refreshToken], properties: { refreshToken: { type: string } } } } }
      responses:
        '200': { description: Yeni cütlük, content: { application/json: { schema: { $ref: '#/components/schemas/TokenResponse' } } } }
        '401': { $ref: '#/components/responses/Problem' }
  /auth/logout:
    post:
      operationId: postAuthLogout
      tags: [Auth]
      summary: Sessiyanı bağla (refresh token family-si ləğv olunur, idempotentdir)
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [refreshToken], properties: { refreshToken: { type: string } } } } }
      responses: { '204': { description: OK } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
...
  responses:
    Problem:
      description: Xəta (RFC 9457)
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
  schemas:
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string, format: uri-reference, description: 'https://errors.dentacore.app/<code>' }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        code: { type: string, examples: [appointment.overlap, auth.invalid_credentials, billing.overpayment] }
        traceId: { type: string }
        errors: { type: object, additionalProperties: { type: array, items: { type: string } } }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, format: password }
        deviceName: { type: string }
        deviceId: { type: string, format: uuid }
    TokenResponse:
      type: object
      required: [accessToken, refreshToken, expiresIn]
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        expiresIn: { type: integer, description: Saniyə (600) }
        tokenType: { type: string, const: Bearer }
    TwoFactorChallenge:
      type: object
      properties: { challengeId: { type: string }, methods: { type: array, items: { type: string, enum: [totp, webauthn] } } }
...